checking if file exists before backup

+++ ksu-so-scrapper.js
@@ -2,7 +2,9 @@ var request = require('request'),
 	cheerio = require('cheerio'),
 	fs = require('fs');
 
-fs.renameSync('organisations.csv', 'organisations.csv.bak');
+if (fs.existsSync('organisations.csv')) {
+  fs.renameSync('organisations.csv', 'organisations.csv.bak');
+}
 
 request({
 	uri: 'http://ksu.org.mt/index.php/representation/student-organisations'
